[NotoSansMono-MM.glyphs] drop kerning to fix variable build
Running fontmake to build a variable Noto Sans Mono fails with

AssertionError: ((2, [3, 2, 2, 2, 3, 2, 2, 2]), 'int', '.FeatureCount', 'FeatureList', '.FeatureList', 'GPOS', '.table', 'table_G_P_O_S_')

This is because only in a couple of masters, there are some a few
kerning pairs, probably some unwanted leftovers since this is a
monospaced font so one would expect no kerning as the advance widths
are all the same.

The presence of the kerning pairs makes ufo2ft build a kern feature
for those masters, and then varLib fails because it doesn't know
how to handle this incompatible GPOS tables.

If I remove the kerning, I can successfully build the variable font.