NotoSerif-ItalicMM.glyphs: make 'dieresiscaron' glyph interpolatable
cu2qu raises an IncompatibleGlyphsError because the CondensedBoldItalic has a line whereas all the other masters have a curve.
Note that this glyph's Export flag is marked as false, but cu2qu still attempts to convert all the glyphs in the fonts, even those that according to Glyphs.app should not be exported in the final font. TODO: do someting about it