NotoSans-MM.glyphs: remove invalid kern pair
the latest ufo2ft (pre-release v2.8.0b1) checks if any kern pair exists between a base glyph and a combining mark glyph (per GDEF GlyphClassDef). If any is found, it splits the auto-generated kern lookup into two, one for regular kern pairs with `lookupflag ignoreMarks` set, the other lookup with pairs between base and mark glyphs with ignoreMarks flag unset (this is to support Noto Siddham specifically).

However, attempting to build NotoSans-MM.glyphs into a variable font using this version of ufo2ft produces an error because the CondensedBold master (only, not the rest of the masters) contains a kern pair involving a combining mark (doublebrevecomb). This triggers the MarkFeatureWriter to generate an extra lookup for the CondensedBold, but then fonttools varLib fails to merge the masters' GPOS tables because they no longer align (one master has one additional lookup).

This patch gets rid of the offending kern pair in NotoSans-MM CondensedBold master.
Note this pair has 0 value, so it was not actually doing anything.